+++ src/sre_agent/cli/mode/remote/__init__.py
@@ -1,0 +1,1 @@
+"""Remote mode package."""
