+++ src/sre_agent/cli/mode/remote/aws/__init__.py
@@ -1,0 +1,1 @@
+"""AWS remote deployment package."""
