+++ src/sre_agent/cli/mode/remote/aws/ecs/__init__.py
@@ -1,0 +1,1 @@
+"""AWS ECS remote deployment package."""
