+++ src/sre_agent/config/__init__.py
@@ -1,0 +1,1 @@
+"""Shared configuration helpers."""
