+++ src/sre_agent/core/agent.py
@@ -2,9 +2,9 @@
 
 from pydantic_ai import Agent
 
-from sre_agent.core.config import AgentConfig, get_config
 from sre_agent.core.models import ErrorDiagnosis
 from sre_agent.core.prompts import SYSTEM_PROMPT, build_diagnosis_prompt
+from sre_agent.core.settings import AgentSettings, get_settings
 from sre_agent.core.tools import (
     create_cloudwatch_toolset,
     create_github_mcp_toolset,
@@ -12,17 +12,17 @@ from sre_agent.core.tools import (
 )
 
 
-def create_sre_agent(config: AgentConfig | None = None) -> Agent[None, ErrorDiagnosis]:
+def create_sre_agent(config: AgentSettings | None = None) -> Agent[None, ErrorDiagnosis]:
     """Create the SRE Agent with all toolsets configured.
 
     Args:
-        config: Optional AgentConfig. If not provided, loads from environment.
+        config: Optional AgentSettings. If not provided, loads from environment.
 
     Returns:
         Configured pydantic-ai Agent with structured output.
     """
     if config is None:
-        config = get_config()
+        config = get_settings()
 
     toolsets = [
         create_cloudwatch_toolset(config),
@@ -42,7 +42,7 @@ async def diagnose_error(
     log_group: str,
     service_name: str,
     time_range_minutes: int = 10,
-    config: AgentConfig | None = None,
+    config: AgentSettings | None = None,
 ) -> ErrorDiagnosis:
     """Run a diagnosis for errors in a specific log group.
 
@@ -56,7 +56,7 @@ async def diagnose_error(
         ErrorDiagnosis with findings and suggested fixes.
     """
     if config is None:
-        config = get_config()
+        config = get_settings()
 
     agent = create_sre_agent(config)
     prompt = build_diagnosis_prompt(config, log_group, service_name, time_range_minutes)
